presintaion worthy

+++ promoto/backend/backend/events/models.py
@@ -6,7 +6,7 @@ import datetime
 class Event(models.Model):
     title = models.CharField(max_length=225)
     location = models.CharField(max_length=225)
-    ownerId = models.BigIntegerField()
+    ownerId = models.IntegerField()
     date = models.DateField(datetime.datetime.now)
     group = models.CharField(max_length=225, null="misc")
     coverImg = models.CharField(max_length=225, default="test.png")
